requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
django-extensions==2.2.5
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.10.3
Faker==2.0.3
gunicorn==19.9.0
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
whitenoise==4.1.4
